fix: change file strcuture and fix init function
