+++ src/app/app.routes.ts
@@ -4,7 +4,7 @@ import { NotFoundComponent } from './pages/404-page.component';
 import { LoginComponent } from './pages/login.component';
 import { RegistrationComponent } from './pages/registration/registration.component';
 import { DetailedProductInformationComponent } from './pages/detailed-product-information.component';
-import { ProfileComponent } from './pages/profile.component';
+import { ProfileComponent } from './pages/user-profile/user-profile.component';
 
 export const routes: Routes = [
   {
