+++ src/app/pages/user-profile/user-profile.component.html
@@ -1,1 +1,7 @@
-<p>user-profile works!</p>
+<div>
+  <h1>
+    Here should be User Profile Page. It should display information about a
+    user's profile.
+  </h1>
+  <button (click)="goBack()">Back</button>
+</div>
